move require inference to seperate file
