+++ type-checker/esprima-verifiers/call-expression.js
@@ -5,6 +5,7 @@ var series = require('run-series');
 var console = require('console');
 
 var verify = require('../verify-esprima-ast.js');
+var inferTypeFromRequire = require('../infer-type-from-require.js');
 var typeCheck = require('../../type-checker/');
 var checkSubType = require('../check-sub-type.js');
 
@@ -63,38 +64,6 @@ function callExpression(node, meta, callback) {
         // loading the source and analyzing it either by loading
         // the correct jsig definition file or
         // by doing type inference
-        getASTForRequire(node, meta, callback);
+        inferTypeFromRequire(node, meta, callback);
     }
-}
-
-function getASTForRequire(node, meta, callback) {
-    var arg = node.arguments[0].value;
-
-    var dirname = path.dirname(meta.filename);
-    var uri;
-    // TODO handle non local uris
-    // TODO replace resolve logic with node-resolve module
-    if (arg[0] === '.') {
-        uri = path.resolve(dirname, arg);
-    }
-
-    if (!uri) {
-        console.log('arg', node);
-        console.warn('skipping require analysis for', arg);
-        return callback(null);
-    }
-
-    // handle folders. lengthen to index.js
-    if (uri.substr(-3) !== '.js') {
-        uri = path.join(uri, 'index.js');
-    }
-
-    typeCheck(uri, function onChecked(err2, fileMeta) {
-        if (err2) {
-            return callback(err2);
-        }
-
-        // console.log('requireMeta', meta)
-        callback(null, fileMeta.moduleExportsType);
-    });
 }
